requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0
pyyaml>=6.0.0
tenacity>=8.0.0
aiohttp-sse>=2.1.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-aiohttp>=1.0.0
aioresponses>=0.7.0
coverage>=6.0.0
backoff>=2.2.0
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0 